Adding support for external dispatchers
External dispatchers are useful when you want to switch on a class hierarchy from a library or from generated Java code, which you can't annotate and enrich with class kinds. It allows one to directly use some external function (the dispatcher) which associates an enum constant to every instance of the base class. The dispatcher (instead of the base class) must then be annotated with @Hierarchy("") (an unmatched exception can be specified, but a field spec will be ignored).